Workaround for herbstluftwm setting the cursor on startup
herbstluftwm sets the default cursor for the user when starting up.
However, this clobbers the cursor that was previously set.

This change fetches the xsetroot command that home-manager places into
the ~/.xsession file, and if found, runs it, thus restoring the cursor
to the theme used before herbstluftwm changed it.